requirements.txt
colorama==0.4.6
python-vlc==3.0.20123
toasts-winrt==1.0.0
Windows-Toasts==1.0.2
